[IMP] base_ubl: Add support for ubl-2.2
This adds the schema for the UBL version 2.2

The files have been linted, but...

The linter only complains about some end of lines and end of files
chars.
Also I don't think it is a good idea to change anything in those
files.
But the only way I know about preventing the linter to access them is to
add some exclude statment in the `.pre-commit-config.yaml` file, and
I guess those changes could be overwritten during next OCA pre-commit
changes.