[13.0][FIX] account_invoice_ubl customer ref
From the little I understand of the specification, it seems that the
OrderReference can contain more than an ID element. See :
https://docs.oasis-open.org/ubl/csprd02-UBL-2.3/mod/summary/reports/UBL-Invoice-2.3.html#Table-OrderReference.Details

But this does not pass the validation and it seems that the only
accepted element in OrderReference is ID.

The goal of this PR is to add the customer reference and ID of the
sale.order being invoiced.