Add 2 hooks to get invoice content and partner
This allow to give an other mean of import of invoice content than from a file.
Which is useful for QR scan in case of l10n_ch_qr_bill_scan module.

Plus add a hook when getting the partner that can accept an action.
This action can be a new wizard step to define the partner.
Also a hook needed for l10n_ch_qr_bill_scan.