account_invoice_import: FIX import config creation from partner
When you import an invoice via the wizard for a new partner that doesn't
exist yet, the wizard will propose you to create a new one. Once the
partner is saved, when you click on "0 invoice import configs", you
would get a bad value in the partner_id field. This commit fixes it.