[FIX] pin PyPDF2 dependency.
PyPDF2 dependency is already included in odoo CE requirements.

However, it is recommended by the repository mantainers to specify
a range of versions that is valid for the correct behavior of the
modules directly using it.

In this case, any version prior to 2.0.0 is valid, and this
also meets standard odoo requirement (1.26.0).