edi: drop ack record auto create
Creating the ack was kind of... an hack :)

When I added this behavior we didn't have yet the storage machinery
that looks automatically for a new file.
Also for webservices seems not relevant since we can configure endpoints
that can create the ack when needed.

All in all, I don't think this feature is needed.
Worse, it adds some overhead when a record gets created
and can generate write conflicts when exchanges are processed immediately
overlapping with the creation of the ack record.