edi: fix _get_component ctx propagation
The env_ctx that might come from a exchange type settings
was not propagated to the exchange record
which is made directly available to the work ctx of the component.

Before this change, if you had any method on the exchange record
or on the related record relying on a ctx key,
the ctx key was not propagated to all records' context.