account_invoice_import_simple_pdf: remove pdfplumber
pdfplumber depends on Pillow >= 9.1, whereas Odoo v14 depends on
Pillow==8.1.1. This causes python dependency issues. So, to make it
easier to deploy the module account_invoice_import_simple_pdf, I decided
to remove the support for pdfplumber.
Since october 2023, we have support for pypdf, which is also
a pure-python lib and doesn't depend on additional python libs, so it's
kind of a replacement for pdfplumber.