[FIX] base_import_pdf_by_template: Improve the import process to account for fixed values
When fixed values are defined, default_x was used in the context of Form() but in some
cases this was not respected (for example if the value was changed in an onchange, warehouse_id
field of sale orders).
We changed the way of creating the records to always take into account the fixed values.

TT50650

[BOT] post-merge updates